Moving tests to test folder
