+++ src/client/src/.tests/components/App.test.js
@@ -2,9 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import {shallow, mount, configure} from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16'
-import App from './App';
+import App from '../../App';
 import {BrowserRouter as Router, Route} from 'react-router'
-import {expect} from 'chai'
 
 configure({adapter: new Adapter()})
